Set framerate & pts when yuv piping
